Restructure into a proper module
