+++ examples/running_mean/standard_scaler/error.rs
@@ -1,0 +1,7 @@
+use std::error::Error;
+
+/// Fast-and-dirty error struct
+#[derive(Debug, Eq, PartialEq, From, Display)]
+pub struct ScalingError {}
+
+impl Error for ScalingError {}
